src/requirements.txt
# ─── Web Framework ────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6

# ─── Database ─────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.11

# ─── MongoDB ──────────────────────────────────
motor==3.6.0

# ─── Supabase ─────────────────────────────────
supabase==2.10.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
cryptography==42.0.5

# ─── Config ───────────────────────────────────
pydantic-settings==2.6.1
pydantic[email]==2.10.3
python-dotenv==1.0.1

# ─── HTTP Client ──────────────────────────────
httpx==0.27.2

# ─── Geometry ─────────────────────────────────
shapely==2.0.6

# ─── Admin Panel ──────────────────────────────
sqladmin==0.19.0
itsdangerous==2.2.0

# ─── Push Notifications ───────────────────────
firebase-admin==6.6.0

# ─── Payments ─────────────────────────────────
razorpay==1.4.1

# ─── Rate Limiting ────────────────────────────
slowapi==0.1.9

# ─── Logging ──────────────────────────────────
structlog==24.4.0

# ─── Utils ────────────────────────────────────
python-multipart==0.0.12
passlib[bcrypt]==1.7.4

# ─── Testing ──────────────────────────────────
pytest==8.3.2
pytest-asyncio==0.23.8